rename and fix econ_miner
